adjust payment method logic
