+++ clients/apps/web/src/components/CreditCardBrandIcon/index.tsx
@@ -38,7 +38,11 @@ const CreditCardBrandIcon = ({
 }: SVGProps<SVGSVGElement> & { brand: string }) => {
   const Icon = getCardIcon(brand)
   return (
-    <svg width="1.5em" height="1em" viewBox="0 0 24 16" {...props}>
+    <svg
+      className="dark:border-polar-700 rounded-lg border border-gray-200 p-2"
+      viewBox="0 0 24 16"
+      {...props}
+    >
       {Icon}
     </svg>
   )
