+++ clients/apps/web/src/components/CustomerPortal/CustomerPortalOverview.tsx
@@ -1,8 +1,14 @@
 import revalidate from '@/app/actions'
+import { useCustomerPaymentMethods } from '@/hooks/queries'
 import { Client, schemas } from '@polar-sh/client'
+import Button from '@polar-sh/ui/components/atoms/Button'
 import { useCallback } from 'react'
+import { Modal } from '../Modal'
+import { useModal } from '../Modal/useModal'
+import { Well, WellContent, WellFooter, WellHeader } from '../Shared/Well'
 import CustomerSubscriptionDetails from '../Subscriptions/CustomerSubscriptionDetails'
-import { CustomerUsage } from './CustomerUsage'
+import { AddPaymentMethodModal } from './AddPaymentMethodModal'
+import PaymentMethod from './PaymentMethod'
 
 interface CustomerPortalOverviewProps {
   organization: schemas['Organization']
@@ -17,13 +23,39 @@ export const CustomerPortalOverview = ({
   products,
   api,
 }: CustomerPortalOverviewProps) => {
+  const {
+    isShown: isAddPaymentMethodModalOpen,
+    hide: hideAddPaymentMethodModal,
+    show: showAddPaymentMethodModal,
+  } = useModal()
+
   const multipleSubscriptions =
     organization.subscription_settings.allow_multiple_subscriptions
 
   const activeSubscription = subscriptions.find((s) => s.status === 'active')
 
+  const { data: paymentMethods } = useCustomerPaymentMethods(api)
+
   return (
     <div className="flex flex-col gap-y-8">
+      <Well className="flex flex-col gap-y-4">
+        <WellHeader>
+          <h3 className="text-xl">Payment Method</h3>
+          <p className="dark:text-polar-500 text-sm text-gray-500">
+            The method used for subscriptions & one-time purchases
+          </p>
+        </WellHeader>
+        <WellContent>
+          {paymentMethods?.items.map((pm) => (
+            <PaymentMethod paymentMethod={pm} api={api} />
+          ))}
+        </WellContent>
+        <WellFooter>
+          <Button className="self-start" onClick={showAddPaymentMethodModal}>
+            Add Payment Method
+          </Button>
+        </WellFooter>
+      </Well>
       {multipleSubscriptions ? (
         <MultipleSubscriptionOverview
           api={api}
@@ -39,82 +71,20 @@ export const CustomerPortalOverview = ({
           products={products}
         />
       )}
-
-      <div className="dark:bg-polar-900 rounded-xl bg-white">
-        <div className="flex flex-col">
-          <div className="dark:border-polar-700 flex items-center justify-between border-b p-4">
-            <h2 className="text-lg font-medium">Overview</h2>
-            <div className="flex items-center gap-2">
-              <span className="dark:text-polar-500 text-sm text-gray-500">
-                Current Billing Cycle
-              </span>
-              <span className="text-sm">Jan 27, 9:00 - Feb 27, 9:00</span>
-            </div>
-          </div>
-
-          <div className="p-4">
-            <div className="grid grid-cols-[1fr,100px,100px,100px] gap-4">
-              <div className="dark:text-polar-500 text-sm text-gray-500">
-                Product
-              </div>
-              <div className="dark:text-polar-500 text-sm text-gray-500">
-                Included
-              </div>
-              <div className="dark:text-polar-500 text-sm text-gray-500">
-                On-demand
-              </div>
-              <div className="dark:text-polar-500 text-sm text-gray-500">
-                Charge
-              </div>
-
-              {[
-                {
-                  name: 'Observability Events',
-                  included: '1M / 1M',
-                  onDemand: '~15M',
-                  charge: '$13.25',
-                },
-                {
-                  name: 'Edge Middleware Invocations',
-                  included: '1M / 1M',
-                  onDemand: '~2.7M',
-                  charge: '$2.60',
-                },
-                {
-                  name: 'Function Invocations',
-                  included: '1M / 1M',
-                  onDemand: '~2.3M',
-                  charge: '$1.80',
-                },
-                {
-                  name: 'Edge Function Execution Units',
-                  included: '125k / 1M',
-                  onDemand: '$0',
-                  charge: '$0',
-                },
-                {
-                  name: 'Edge Requests',
-                  included: '10M / 1M',
-                  onDemand: '$0',
-                  charge: '$0',
-                },
-              ].map((item) => (
-                <div key={item.name}>
-                  <div className="flex items-center gap-2">
-                    <div className="h-2 w-2 rounded-full bg-blue-500" />
-                    <span>{item.name}</span>
-                  </div>
-                  <div>{item.included}</div>
-                  <div>{item.onDemand}</div>
-                  <div className="text-green-500">{item.charge}</div>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <CustomerUsage organizationId={organization.id} />
+      <Modal
+        isShown={isAddPaymentMethodModalOpen}
+        hide={hideAddPaymentMethodModal}
+        modalContent={
+          <AddPaymentMethodModal
+            api={api}
+            onPaymentMethodAdded={() => {
+              revalidate(`customer_portal`)
+              hideAddPaymentMethodModal()
+            }}
+            hide={hideAddPaymentMethodModal}
+          />
+        }
+      />
     </div>
   )
 }
@@ -136,8 +106,8 @@ const SingleSubscriptionOverview = ({
   }, [])
 
   return (
-    <div className="flex flex-col gap-y-4">
-      <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
+    <div className="flex flex-col gap-y-6">
+      <div className="grid grid-cols-1 gap-6 md:grid-cols-2">
         {subscription ? (
           <CustomerSubscriptionDetails
             api={api}
